services/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose[cryptography]
email-validator
